Enhance extraction file
Created a new python file to call into main python extraction file make the connection to the database code even more reusable
