+++ Extract_Script_CSV.py
@@ -2,8 +2,8 @@
 import pandas as pd
 import requests
 from io import StringIO
-import psycopg2
-import os
+from init import ConnectionToDatabase
+import os 
 
 ## Read the csv file 
 url = "https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/owid-covid-data.csv"
@@ -28,23 +28,6 @@ host = 'ep-lively-term-abff86z3-pooler.eu-west-2.aws.neon.tech'
 port = '5432'
 database = 'neondb'
 sslmode = 'require'  # Neon requires SSL
-
-def ConnectionToDatabase(username,password,host,port,database,sslmode):
-
-## Test connection  
-    try:
-        connection = psycopg2.connect(
-            host=host,
-            database=database,
-            user=username,
-            password=password,
-            port=port,
-            sslmode=sslmode
-        )
-        print("Successfully connected to Neon PostgreSQL!")
-        connection.close()
-    except Exception as e:
-        print(" Connection failed:", e)
 
 ## run the connection function 
 ConnectionToDatabase(username,password,host,port,database,sslmode)
